+++ core/transaction/create_multisig.go
@@ -52,7 +52,7 @@ func (data CreateMultisigData) Gas() int64 {
 	return commissions.CreateMultisig
 }
 
-func (data CreateMultisigData) Run(tx *Transaction, context *state.StateDB, isCheck bool, rewardPool *big.Int, currentBlock int64) Response {
+func (data CreateMultisigData) Run(tx *Transaction, context *state.StateDB, isCheck bool, rewardPool *big.Int, currentBlock uint64) Response {
 	sender, _ := tx.Sender()
 
 	response := data.BasicCheck(tx, context)