+++ core/transaction/edit_candidate.go
@@ -44,7 +44,7 @@ func (data EditCandidateData) Gas() int64 {
 	return commissions.EditCandidate
 }
 
-func (data EditCandidateData) Run(tx *Transaction, context *state.StateDB, isCheck bool, rewardPool *big.Int, currentBlock int64) Response {
+func (data EditCandidateData) Run(tx *Transaction, context *state.StateDB, isCheck bool, rewardPool *big.Int, currentBlock uint64) Response {
 	sender, _ := tx.Sender()
 
 	response := data.BasicCheck(tx, context)