+++ core/transaction/redeem_check.go
@@ -48,7 +48,7 @@ func (data RedeemCheckData) Gas() int64 {
 	return commissions.RedeemCheckTx
 }
 
-func (data RedeemCheckData) Run(tx *Transaction, context *state.StateDB, isCheck bool, rewardPool *big.Int, currentBlock int64) Response {
+func (data RedeemCheckData) Run(tx *Transaction, context *state.StateDB, isCheck bool, rewardPool *big.Int, currentBlock uint64) Response {
 	sender, _ := tx.Sender()
 
 	response := data.BasicCheck(tx, context)