+++ CHANGELOG.md
@@ -11,6 +11,7 @@ IMPROVEMENT
 - [core] Delete coins with 0 reserves #217
 - [genesis] Add option to export/import state
 - [api] Add ?include_stakes to /candidates endpoint #222
+- [events] Add CoinLiquidation event #221
 
 BUG FIXES